nest mind/requirements.txt
streamlit==1.40.0
flask==3.0.3
google-generativeai==0.5.2
python-dotenv==1.0.1
tenacity==9.0.0
rich==13.7.1
requests==2.32.3